Added multiple things
added diverge, fortified plating, shield, changed damage, defense, health and speed to float.

+++ surrounded/Assets/Scripts/GameControl/WorldBoundry.cs
@@ -13,7 +13,7 @@ public class WorldBoundary : MonoBehaviour
         if (other.gameObject.CompareTag("Player"))
         {
             PlayerController target = other.gameObject.GetComponent<PlayerController>();
-            int dmg = (target.health)/3;
+            float dmg = (target.health)/3;
             dot = StartCoroutine(DOT(target,dmg));
          }
         if (other.gameObject.CompareTag("Enemy"))
@@ -25,7 +25,7 @@ public class WorldBoundary : MonoBehaviour
     public void OnCollisionExit2D(Collision2D other){
         StopCoroutine(dot);
     }
-    private IEnumerator DOT(PlayerController target, int damage){
+    private IEnumerator DOT(PlayerController target, float damage){
         while(target != null){
             target.takeDamage(damage);
             yield return new WaitForSeconds(1);
